requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
djangorestframework==3.13.1
idna==3.3
Pillow==9.2.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
transliterate==1.10.2
tzdata==2022.1
urllib3==1.26.11
